+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/instrumentation/fakes/TestSpanAddingParameters.kt
@@ -47,4 +47,7 @@ class TestSpanAddingParameters(
 
   override val appStartEnabled: Property<Boolean>
     get() = TODO()
+
+  override val binderIpcEnabled: Property<Boolean>
+    get() = TODO()
 }